Refactor Gear Math + Add AEM CAN Wideband (#15)
* Refactor Gear Math + Add AEM CAN Wideband

Tire size is now a constant that can be modified and Gauge.S will figure out gear from that.

Also added parameter for AEM 30-0310 CAN Wideband O2 sensor. This is disabled by default as it is not a standard sensor on the car.

* Refactor Gear Math + Add AEM CAN Wideband

Move AFR logic to it's own file.